refactor: split host network routes
